+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - A join between triples maps with different logical targets might result in quads ending up in wrong targets.
 - rmle:contentTypeHeader and rmle:acceptHeader is replaced by generic htv:headers for Http Request Targets.
 - Comparing htv:headers might go wrong due to String comparison.
+- Conversion to new RML goes wrong when an object map has term type Literal.
 
 ### Added
 - Option to serialize output as Jelly.